+++ config/aggregate_test.go
@@ -17,9 +17,9 @@ func TestAggregate(t *testing.T) {
 	testHandler(
 		t,
 		configbuilder.Aggregate,
-		runtimeconfig.FromAggregate,
-		func(fn func(dogma.AggregateConfigurer)) dogma.AggregateMessageHandler {
-			return &AggregateMessageHandlerStub{ConfigureFunc: fn}
+		runtimeconfig.FromAggregate[dogma.AggregateRoot],
+		func(fn func(dogma.AggregateConfigurer)) dogma.AggregateMessageHandler[dogma.AggregateRoot] {
+			return dogma.UntypedAggregateMessageHandler(&AggregateMessageHandlerStub[*AggregateRootStub]{ConfigureFunc: fn})
 		},
 	)
 
@@ -29,7 +29,7 @@ func TestAggregate(t *testing.T) {
 			{
 				Name:  "valid",
 				Error: ``,
-				Component: runtimeconfig.FromAggregate(&AggregateMessageHandlerStub{
+				Component: runtimeconfig.FromAggregate(&AggregateMessageHandlerStub[*AggregateRootStub]{
 					ConfigureFunc: func(c dogma.AggregateConfigurer) {
 						c.Identity("name", "19cb98d5-dd17-4daf-ae00-1b413b7b899a")
 						c.Routes(
@@ -70,7 +70,7 @@ func TestAggregate(t *testing.T) {
 				Name: "no runtime values with ForExecution() option",
 				Error: multiline(
 					`aggregate:SomeAggregate is invalid:`,
-					`  - dogma.AggregateMessageHandler value is unavailable`,
+					`  - dogma.AggregateMessageHandler[github.com/dogmatiq/dogma.AggregateRoot] value is unavailable`,
 					`  - route:handles-command:SomeCommand is invalid:`,
 					`      - message type ID is unavailable`,
 					`      - message type is unavailable`,
@@ -114,22 +114,22 @@ func TestAggregate(t *testing.T) {
 					`  - no handles-command routes`,
 					`  - no records-event routes`,
 				),
-				Component: runtimeconfig.FromAggregate(nil),
+				Component: runtimeconfig.FromAggregate[*AggregateRootStub](nil),
 			},
 			{
 				Name: "unconfigured aggregate",
 				Error: multiline(
-					`aggregate:AggregateMessageHandlerStub is invalid:`,
+					`aggregate:AggregateMessageHandlerStub[AggregateRootStub] is invalid:`,
 					`  - no identity`,
 					`  - no handles-command routes`,
 					`  - no records-event routes`,
 				),
-				Component: runtimeconfig.FromAggregate(&AggregateMessageHandlerStub{}),
+				Component: runtimeconfig.FromAggregate(&AggregateMessageHandlerStub[*AggregateRootStub]{}),
 			},
 			{
 				Name:  "aggregate identity must be valid",
-				Error: `aggregate:AggregateMessageHandlerStub is invalid: identity:name/non-uuid is invalid: invalid key ("non-uuid"), expected an RFC 4122/9562 UUID`,
-				Component: runtimeconfig.FromAggregate(&AggregateMessageHandlerStub{
+				Error: `aggregate:AggregateMessageHandlerStub[AggregateRootStub] is invalid: identity:name/non-uuid is invalid: invalid key ("non-uuid"), expected an RFC 4122/9562 UUID`,
+				Component: runtimeconfig.FromAggregate(&AggregateMessageHandlerStub[*AggregateRootStub]{
 					ConfigureFunc: func(c dogma.AggregateConfigurer) {
 						c.Identity("name", "non-uuid")
 						c.Routes(
@@ -141,8 +141,8 @@ func TestAggregate(t *testing.T) {
 			},
 			{
 				Name:  "aggregate must not have multiple identities",
-				Error: `aggregate:AggregateMessageHandlerStub is invalid: multiple identities`,
-				Component: runtimeconfig.FromAggregate(&AggregateMessageHandlerStub{
+				Error: `aggregate:AggregateMessageHandlerStub[AggregateRootStub] is invalid: multiple identities`,
+				Component: runtimeconfig.FromAggregate(&AggregateMessageHandlerStub[*AggregateRootStub]{
 					ConfigureFunc: func(c dogma.AggregateConfigurer) {
 						c.Identity("foo", "63bd2756-2397-4cae-b33b-96e809b384d8")
 						c.Identity("foo", "63bd2756-2397-4cae-b33b-96e809b384d8")
@@ -156,8 +156,8 @@ func TestAggregate(t *testing.T) {
 			},
 			{
 				Name:  "aggregate must handle at least one command type",
-				Error: `aggregate:AggregateMessageHandlerStub is invalid: no handles-command routes`,
-				Component: runtimeconfig.FromAggregate(&AggregateMessageHandlerStub{
+				Error: `aggregate:AggregateMessageHandlerStub[AggregateRootStub] is invalid: no handles-command routes`,
+				Component: runtimeconfig.FromAggregate(&AggregateMessageHandlerStub[*AggregateRootStub]{
 					ConfigureFunc: func(c dogma.AggregateConfigurer) {
 						c.Identity("handler", "d1e04684-ec56-44a7-8c7d-f111b2d6b2d2")
 						c.Routes(
@@ -169,8 +169,8 @@ func TestAggregate(t *testing.T) {
 			},
 			{
 				Name:  "aggregate must record at least one event type",
-				Error: `aggregate:AggregateMessageHandlerStub is invalid: no records-event routes`,
-				Component: runtimeconfig.FromAggregate(&AggregateMessageHandlerStub{
+				Error: `aggregate:AggregateMessageHandlerStub[AggregateRootStub] is invalid: no records-event routes`,
+				Component: runtimeconfig.FromAggregate(&AggregateMessageHandlerStub[*AggregateRootStub]{
 					ConfigureFunc: func(c dogma.AggregateConfigurer) {
 						c.Identity("handler", "d1e04684-ec56-44a7-8c7d-f111b2d6b2d2")
 						c.Routes(
@@ -182,8 +182,8 @@ func TestAggregate(t *testing.T) {
 			},
 			{
 				Name:  "aggregate must not have multiple routes for the same command type",
-				Error: `aggregate:AggregateMessageHandlerStub is invalid: multiple handles-command routes for *github.com/dogmatiq/enginekit/enginetest/stubs.CommandStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]`,
-				Component: runtimeconfig.FromAggregate(&AggregateMessageHandlerStub{
+				Error: `aggregate:AggregateMessageHandlerStub[AggregateRootStub] is invalid: multiple handles-command routes for *github.com/dogmatiq/enginekit/enginetest/stubs.CommandStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]`,
+				Component: runtimeconfig.FromAggregate(&AggregateMessageHandlerStub[*AggregateRootStub]{
 					ConfigureFunc: func(c dogma.AggregateConfigurer) {
 						c.Identity("handler", "d1e04684-ec56-44a7-8c7d-f111b2d6b2d2")
 						c.Routes(
@@ -196,8 +196,8 @@ func TestAggregate(t *testing.T) {
 			},
 			{
 				Name:  "aggregate must not have multiple routes for the same event type",
-				Error: `aggregate:AggregateMessageHandlerStub is invalid: multiple records-event routes for *github.com/dogmatiq/enginekit/enginetest/stubs.EventStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]`,
-				Component: runtimeconfig.FromAggregate(&AggregateMessageHandlerStub{
+				Error: `aggregate:AggregateMessageHandlerStub[AggregateRootStub] is invalid: multiple records-event routes for *github.com/dogmatiq/enginekit/enginetest/stubs.EventStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]`,
+				Component: runtimeconfig.FromAggregate(&AggregateMessageHandlerStub[*AggregateRootStub]{
 					ConfigureFunc: func(c dogma.AggregateConfigurer) {
 						c.Identity("handler", "d1e04684-ec56-44a7-8c7d-f111b2d6b2d2")
 						c.Routes(
@@ -249,14 +249,14 @@ func TestAggregate(t *testing.T) {
 		describeTestCases{
 			{
 				Name:   "complete",
-				String: `aggregate:AggregateMessageHandlerStub`,
+				String: `aggregate:AggregateMessageHandlerStub[AggregateRootStub]`,
 				Description: multiline(
-					`valid aggregate *github.com/dogmatiq/enginekit/enginetest/stubs.AggregateMessageHandlerStub`,
+					`valid aggregate *github.com/dogmatiq/enginekit/enginetest/stubs.AggregateMessageHandlerStub[*github.com/dogmatiq/enginekit/enginetest/stubs.AggregateRootStub]`,
 					`  - valid identity name/19cb98d5-dd17-4daf-ae00-1b413b7b899a`,
 					`  - valid handles-command route for *github.com/dogmatiq/enginekit/enginetest/stubs.CommandStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]`,
 					`  - valid records-event route for *github.com/dogmatiq/enginekit/enginetest/stubs.EventStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]`,
 				),
-				Component: runtimeconfig.FromAggregate(&AggregateMessageHandlerStub{
+				Component: runtimeconfig.FromAggregate(&AggregateMessageHandlerStub[*AggregateRootStub]{
 					ConfigureFunc: func(c dogma.AggregateConfigurer) {
 						c.Identity("name", "19cb98d5-dd17-4daf-ae00-1b413b7b899a")
 						c.Routes(
@@ -268,15 +268,15 @@ func TestAggregate(t *testing.T) {
 			},
 			{
 				Name:   "disabled",
-				String: `aggregate:AggregateMessageHandlerStub`,
+				String: `aggregate:AggregateMessageHandlerStub[AggregateRootStub]`,
 				Description: multiline(
-					`valid aggregate *github.com/dogmatiq/enginekit/enginetest/stubs.AggregateMessageHandlerStub`,
+					`valid aggregate *github.com/dogmatiq/enginekit/enginetest/stubs.AggregateMessageHandlerStub[*github.com/dogmatiq/enginekit/enginetest/stubs.AggregateRootStub]`,
 					`  - valid identity name/19cb98d5-dd17-4daf-ae00-1b413b7b899a`,
 					`  - valid handles-command route for *github.com/dogmatiq/enginekit/enginetest/stubs.CommandStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]`,
 					`  - valid records-event route for *github.com/dogmatiq/enginekit/enginetest/stubs.EventStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]`,
 					`  - valid disabled flag, set to true`,
 				),
-				Component: runtimeconfig.FromAggregate(&AggregateMessageHandlerStub{
+				Component: runtimeconfig.FromAggregate(&AggregateMessageHandlerStub[*AggregateRootStub]{
 					ConfigureFunc: func(c dogma.AggregateConfigurer) {
 						c.Identity("name", "19cb98d5-dd17-4daf-ae00-1b413b7b899a")
 						c.Routes(
@@ -325,14 +325,14 @@ func TestAggregate(t *testing.T) {
 			},
 			{
 				Name:   "invalid",
-				String: `aggregate:AggregateMessageHandlerStub`,
+				String: `aggregate:AggregateMessageHandlerStub[AggregateRootStub]`,
 				Description: multiline(
-					`invalid aggregate *github.com/dogmatiq/enginekit/enginetest/stubs.AggregateMessageHandlerStub`,
+					`invalid aggregate *github.com/dogmatiq/enginekit/enginetest/stubs.AggregateMessageHandlerStub[*github.com/dogmatiq/enginekit/enginetest/stubs.AggregateRootStub]`,
 					`  - no handles-command routes`,
 					`  - no records-event routes`,
 					`  - valid identity name/19cb98d5-dd17-4daf-ae00-1b413b7b899a`,
 				),
-				Component: runtimeconfig.FromAggregate(&AggregateMessageHandlerStub{
+				Component: runtimeconfig.FromAggregate(&AggregateMessageHandlerStub[*AggregateRootStub]{
 					ConfigureFunc: func(c dogma.AggregateConfigurer) {
 						c.Identity("name", "19cb98d5-dd17-4daf-ae00-1b413b7b899a")
 					},
@@ -340,15 +340,15 @@ func TestAggregate(t *testing.T) {
 			},
 			{
 				Name:   "invalid sub-component",
-				String: `aggregate:AggregateMessageHandlerStub`,
+				String: `aggregate:AggregateMessageHandlerStub[AggregateRootStub]`,
 				Description: multiline(
-					`valid aggregate *github.com/dogmatiq/enginekit/enginetest/stubs.AggregateMessageHandlerStub`,
+					`valid aggregate *github.com/dogmatiq/enginekit/enginetest/stubs.AggregateMessageHandlerStub[*github.com/dogmatiq/enginekit/enginetest/stubs.AggregateRootStub]`,
 					`  - invalid identity name/non-uuid`,
 					`      - invalid key ("non-uuid"), expected an RFC 4122/9562 UUID`,
 					`  - valid handles-command route for *github.com/dogmatiq/enginekit/enginetest/stubs.CommandStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]`,
 					`  - valid records-event route for *github.com/dogmatiq/enginekit/enginetest/stubs.EventStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]`,
 				),
-				Component: runtimeconfig.FromAggregate(&AggregateMessageHandlerStub{
+				Component: runtimeconfig.FromAggregate(&AggregateMessageHandlerStub[*AggregateRootStub]{
 					ConfigureFunc: func(c dogma.AggregateConfigurer) {
 						c.Identity("name", "non-uuid")
 						c.Routes(
@@ -363,7 +363,7 @@ func TestAggregate(t *testing.T) {
 
 	t.Run("func RouteSet()", func(t *testing.T) {
 		t.Run("it returns the normalized routes", func(t *testing.T) {
-			h := &AggregateMessageHandlerStub{
+			h := &AggregateMessageHandlerStub[*AggregateRootStub]{
 				ConfigureFunc: func(c dogma.AggregateConfigurer) {
 					c.Identity("name", "19cb98d5-dd17-4daf-ae00-1b413b7b899a")
 					c.Routes(